requirements.txt
discord.py
openpyxl
ast
dotenv